Stopped whiskers on Vacuum Stop and Stopped phasing through walls
